Search bar component, deleted unused pages, added READMEs
