+++ flote/src/app/App.tsx
@@ -1,12 +1,10 @@
+import { Navigate, Outlet, Route, Routes } from "react-router-dom";
 import { useEffect } from "react";
-import { Navigate, Outlet, Routes, Route } from "react-router-dom";
 
 import { socket } from "../socket";
 
-import Landing from "./pages/Landing";
-import SignUp from "./pages/SignUp";
-import SignIn from "./pages/SignIn";
 import AccountHome from "./pages/AccountHome";
+import Landing from "./pages/Landing";
 
 const PrivateRoutes = () => {
   const auth = { token: true };
@@ -21,8 +19,6 @@ export default function App() {
   return (
     <Routes>
       <Route path="/" element={<Landing />} />
-      <Route path="sign-up" element={<SignUp />} />
-      <Route path="sign-in" element={<SignIn />} />
       <Route element={<PrivateRoutes />}>
         <Route path="home" element={<AccountHome />} />
       </Route>
