+++ flote/src/components/atoms/Logo.tsx
@@ -6,7 +6,7 @@ type Props = {
 export default function Logo({ className, onClick }: Props) {
   return (
     <h1
-      className={`font-header font-bold text-white text-center ${className}`}
+      className={`font-header font-bold text-white ${className}`}
       onClick={onClick}
     >
       <span className={onClick ? "cursor-pointer" : ""}>flotE</span>
