+++ server/README.md
@@ -1,0 +1,5 @@
+## Starting the server
+```
+npm run devStart
+```
+The server runs on port 3000.
